Combine GraphImage into LineGraph
